Rename tools/isobuild/import-scanner.js to .ts.
Note: I have made no effort to fix any TypeScript errors or warnings,
yet. This is just a placeholder commit so that I can claim this module
publicly, before anyone else attempts converting it.
